# contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Continua API
  version: 0.1.0
  description: AI Agent Observability Platform API
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: /
    description: Relative path (works with any host)
security:
  - auth0Bearer: []
  - apiKey: []
paths:
...
  /v1/ingest:
    post:
      operationId: ingest
      summary: Ingest traces, spans, and events
      description: |
        Batch ingestion endpoint for traces, spans, and events.
        Supports idempotent processing via batch_key.
        Maximum request body size is 5MB.
      tags: [Ingest]
      security:
        - apiKey: []
      parameters:
        - name: sync
          in: query
          description: If true, wait for processing to complete before returning
          schema:
            type: boolean
            default: false
        - name: X-Continua-Async-Version
          in: header
          description: |
            Optional staged-rollout header for true async ingest.
            `2` enables durable async acceptance and background processing.
            Any other value is rejected with `400 unsupported_async_version`.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRequest'
      responses:
        '200':
          description: Batch processed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '202':
          description: Batch accepted for asynchronous processing
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestResponse'
        '400':
          description: Invalid request body or unsupported async version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '413':
          description: Request body too large (exceeds 5MB limit)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SizeError'
...
  /v1/engine/runs:
    post:
      operationId: startEngineRun
      summary: Start an engine workflow run
      tags: [Engine]
      security:
        - apiKey: []
      parameters:
        - name: X-Continua-Engine-Preview
          in: header
          required: true
          description: Required preview header for mutating engine routes.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineStartRunRequest'
      responses:
        '200':
          description: Engine run started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngineStartRunResponse'
        '400':
          description: Invalid request or missing preview header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Engine API disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Duplicate, in-progress, or conflicting request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /v1/engine/activities/claim:
    post:
      operationId: claimRemoteActivityTasks
      summary: Claim remote engine activity tasks
      tags: [Engine]
      security:
        - apiKey: []
      parameters:
        - name: X-Continua-Engine-Preview
          in: header
          required: true
          description: Required preview header for mutating engine routes.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineRemoteActivityClaimRequest'
      responses:
        '200':
          description: Claimed remote activity tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngineRemoteActivityClaimResponse'
        '400':
          description: Invalid request or missing preview header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /v1/engine/projections/backfill:
    post:
      operationId: backfillEngineProjections
      summary: Preview or request bulk projection repair for engine runs
      description: |
        Identifies engine-backed traces eligible for projection repair and, when
        `dry_run=false`, reuses the existing per-run repair service to request
        catch-up. When `engine_projection_state` is omitted, only `summary_only`
        candidates are considered. Explicit `up_to_date`, `catching_up`, and
        `journal_expired` filters return zero eligible rows because those states
        are not valid backfill targets.
      tags: [Engine]
      security:
        - apiKey: []
      parameters:
        - name: X-Continua-Engine-Preview
          in: header
          required: true
          description: Required preview header for mutating engine routes.
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngineProjectionBackfillRequest'
      responses:
        '200':
          description: Backfill preview or apply result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngineProjectionBackfillResponse'
        '400':
          description: Invalid request body, limit above 100, or missing preview header
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Engine API disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/traces:
    get:
      operationId: listTraces
      summary: List traces
      tags: [Traces]
      parameters:
        - $ref: '#/components/parameters/SelectedProjectId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: session_id
          in: query
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          description: Full-text search query (searches trace name and user_id)
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort traces by started time. Ignored when q is present.
          schema:
            type: string
            enum: [started_at]
        - name: sort_dir
          in: query
          description: Sort direction for started time. Ignored when q is present.
          schema:
            type: string
            enum: [asc, desc]
        - name: status
          in: query
          description: Filter by trace status
          schema:
            type: string
            enum: [running, completed, failed]
        - name: start_time_from
          in: query
          description: Filter traces starting at or after this time
          schema:
            type: string
            format: date-time
        - name: start_time_to
          in: query
          description: Filter traces starting at or before this time
          schema:
            type: string
            format: date-time
        - name: user_id
          in: query
          description: Filter by user ID
          schema:
            type: string
        - name: engine_instance_key
          in: query
          description: Filter by engine instance key
          schema:
            type: string
        - name: engine_run_id
          in: query
          description: Filter by engine run ID
          schema:
            type: string
            format: uuid
        - name: engine_definition_name
          in: query
          description: Filter by engine definition name
          schema:
            type: string
        - name: engine_definition_version
          in: query
          description: Filter by engine definition version
          schema:
            type: string
        - name: engine_run_status
          in: query
          description: Filter by engine run lifecycle status
          schema:
            type: string
            enum: [queued, running, waiting, suspended, completed, failed, cancelled, terminated, continued_as_new]
        - name: engine_parent_run_id
          in: query
          description: Filter by engine parent run ID
          schema:
            type: string
            format: uuid
        - name: engine_root_run_id
          in: query
          description: Filter by engine root run ID
          schema:
            type: string
            format: uuid
        - name: engine_child_key
          in: query
          description: Filter by engine child key
          schema:
            type: string
        - name: engine_child_depth
          in: query
          description: Filter by engine child depth
          schema:
            type: integer
        - name: engine_projection_state
          in: query
          description: Filter by engine projection state
          schema:
            $ref: '#/components/schemas/EngineProjectionState'
        - name: has_errors
          in: query
          description: Filter traces with errors (error_count > 0)
          schema:
            type: boolean
        - name: min_duration_ms
          in: query
          description: Filter traces with duration >= this value in milliseconds
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of traces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceList'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/traces/{id}:
    get:
      operationId: getTrace
      summary: Get a trace by ID
      tags: [Traces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
      responses:
        '200':
          description: Trace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TraceDetail'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/traces/{id}/spans:
    get:
      operationId: listSpansByTrace
      summary: List spans for a trace
      tags: [Spans]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
      responses:
        '200':
          description: List of spans
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SpanList'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/traces/{id}/events:
    get:
      operationId: getTraceEvents
      summary: List timeline events for a trace
      tags: [Traces]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
        - name: after
          in: query
          description: Opaque cursor returned by a previous timeline response
          schema:
            type: string
        - name: limit
          in: query
          description: Maximum number of timeline events to return
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 500
      responses:
        '200':
          description: Timeline events for the trace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TimelineResponse'
        '400':
          description: Invalid cursor or pagination parameter
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions:
    get:
      operationId: listSessions
      summary: List sessions
      tags: [Sessions]
      parameters:
        - $ref: '#/components/parameters/SelectedProjectId'
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
        - name: q
          in: query
          description: Search sessions by external ID or name
          schema:
            type: string
        - name: user_id
          in: query
          description: Filter sessions by exact user ID
          schema:
            type: string
        - name: sort_by
          in: query
          description: Sort sessions by creation time or trace count. Ignored when q is present.
          schema:
            type: string
            enum: [created_at, trace_count]
        - name: sort_dir
          in: query
          description: Sort direction for the selected session sort field. Ignored when q is present.
          schema:
            type: string
            enum: [asc, desc]
      responses:
        '200':
          description: List of sessions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionList'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions/{id}:
    get:
      operationId: getSession
      summary: Get a session by ID
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
      responses:
        '200':
          description: Session details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions/{id}/narrative:
    get:
      operationId: getSessionNarrative
      summary: Get a session narrative by ID
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
      responses:
        '200':
          description: Session narrative details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionNarrativeResponse'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /api/sessions/{id}/compare:
    get:
      operationId: getSessionCompare
      summary: Compare two traces within a session
      tags: [Sessions]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - $ref: '#/components/parameters/SelectedProjectId'
        - name: baseline_trace_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: candidate_trace_id
          in: query
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Session trace comparison details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionCompareResponse'
        '400':
          description: Invalid comparison request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized - missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Session or trace not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Comparison exceeds the supported span or semantic event ceiling
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComparisonTooLargeError'
components:
  parameters:
    SelectedProjectId:
      name: project_id
      in: query
      description: Required when authenticating with an Auth0 bearer token. Ignored when the request is already project-scoped by API key.
      schema:
        type: string
        format: uuid
  securitySchemes:
    apiKey:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication
    auth0Bearer:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Auth0 access token for debugger operator authentication
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    SizeError:
      type: object
      required: [error]
      description: Error response for 413 Payload Too Large
      properties:
        error:
          type: string
          description: Error message describing the size limit violation
    EngineProjectionState:
      type: string
      enum: [up_to_date, catching_up, summary_only, journal_expired]
...
    EngineRemoteActivityClaimRequest:
      type: object
      required: [worker_id, activity_types]
      properties:
        worker_id:
          type: string
          minLength: 1
          maxLength: 128
        activity_types:
          type: array
          minItems: 1
          maxItems: 50
          items:
            type: string
            minLength: 1
            maxLength: 256
        lease_duration:
          type: string
          description: Requested lease duration, e.g. 60s or 5m. The server clamps the effective value.
        max_tasks:
          type: integer
          minimum: 1
          maximum: 50
...
    EngineRemoteActivityClaimResponse:
      type: object
      required: [tasks]
      properties:
        tasks:
          type: array
          items:
            $ref: '#/components/schemas/EngineRemoteActivityTask'
...
    EngineProjectionBackfillRequest:
      type: object
      properties:
        dry_run:
          type: boolean
          default: false
        limit:
          type: integer
          default: 50
          maximum: 100
        older_than:
          type: string
          format: date-time
        engine_instance_key:
          type: string
        engine_definition_name:
          type: string
        engine_run_status:
          $ref: '#/components/schemas/EngineRunStatus'
        engine_projection_state:
          allOf:
            - $ref: '#/components/schemas/EngineProjectionState'
          description: |
            Defaults to `summary_only` when omitted. Explicit `up_to_date`,
            `catching_up`, and `journal_expired` return zero eligible rows.
...
    EngineProjectionBackfillResponse:
      type: object
      required: [dry_run, limit, eligible_count, repair_requested_count, skipped_count, results]
      properties:
        dry_run:
          type: boolean
        limit:
          type: integer
        eligible_count:
          type: integer
        repair_requested_count:
          type: integer
        skipped_count:
          type: integer
        results:
          type: array
          items:
            $ref: '#/components/schemas/EngineProjectionBackfillRunResult'
...
    EngineStartRunRequest:
      type: object
      required: [instance_key, definition_name, definition_version, request_key]
      properties:
        instance_key:
          type: string
        definition_name:
          type: string
        definition_version:
          type: string
        request_key:
          type: string
        input:
          description: Workflow input payload (any valid JSON)
        session:
          $ref: '#/components/schemas/EngineStartSession'
        trace:
          $ref: '#/components/schemas/EngineStartTrace'
    EngineStartRunResponse:
      type: object
      required: [run_id, instance_key, trace_id]
      properties:
        run_id:
          type: string
          format: uuid
        instance_key:
          type: string
        trace_id:
          type: string
...
    TraceList:
      type: object
      required: [traces, total]
      properties:
        traces:
          type: array
          items:
            $ref: '#/components/schemas/Trace'
        total:
          type: integer
...
    SpanList:
      type: object
      required: [spans]
      properties:
        spans:
          type: array
          items:
            $ref: '#/components/schemas/Span'
...
    IngestRequest:
      type: object
      required: [batch_key]
      properties:
        batch_key:
          type: string
          description: Unique identifier for idempotency. Same batch_key returns success without reprocessing.
        traces:
          type: array
          items:
            $ref: '#/components/schemas/IngestTraceInput'
        spans:
          type: array
          items:
            $ref: '#/components/schemas/IngestSpanInput'
        events:
          type: array
          items:
            $ref: '#/components/schemas/IngestEventInput'
...
    IngestResponse:
      type: object
      required: [status, batch_key]
      properties:
        status:
          type: string
          enum: [ok, accepted, duplicate, failed]
          description: Processing status. "accepted" for async mode, "duplicate" indicates the batch was already processed.
        batch_key:
          type: string
        batch_id:
          type: string
          format: uuid
          description: |
            Stable batch identifier returned for sync responses, true-async acceptance,
            and duplicates. It may be omitted for the legacy Stage A fake-async path.
        trace_count:
          type: integer
          format: int32
        span_count:
          type: integer
          format: int32
        event_count:
          type: integer
          format: int32
        accepted_count:
          type: integer
          format: int32
        rejected_count:
          type: integer
          format: int32
        errors:
          type: array
          items:
            type: string
...